Adição de estilização do layout com css

+++ App/View/modulos/pessoa/ListaPessoa.php
@@ -2,30 +2,40 @@
 
 <main>
 
-    <table align="center" cellpadding="10px">
+    <?php if(count($linhas) > 0): ?>    
 
-        <tr>
+        <table align="center" cellpadding="10px">
 
-            <th>ID</th>
-            <th>Nome</th>
-            <th>Email</th>
-            <th>Telefone</th>
-
-        </tr>
-
-        <!-- Laço para percorrer todos os registros vindos do banco de dados -->
-        <?php foreach ($linhas as $item): ?>
             <tr>
-                <td><?= $item->id ?></td>
-                <td><?= $item->nome ?></td>
-                <td><?= $item->email ?></td>
-                <td><?= $item->telefone ?></td>
-                <td><a href="/pessoa/cadastro?id=<?= $item->id ?>" ><button>EDITAR</button></a></td>
-                <td><a href="/pessoa/delete?id=<?= $item->id ?>" ><button>EXCLUIR</button></a></td>
-            </tr>
-        <?php endforeach; ?>
 
-    </table>
+                <th>ID</th>
+                <th>Nome</th>
+                <th>Email</th>
+                <th>Telefone</th>
+                <th>  </th>
+                <th>  </th>
+
+            </tr>
+
+            <!-- Laço para percorrer todos os registros vindos do banco de dados -->
+            <?php foreach ($linhas as $item): ?>
+                <tr>
+                    <td><?= $item->id ?></td>
+                    <td><?= $item->nome ?></td>
+                    <td><?= $item->email ?></td>
+                    <td><?= $item->telefone ?></td>
+                    <td><a href="/pessoa/cadastro?id=<?= $item->id ?>" ><button>EDITAR</button></a></td>
+                    <td><a href="/pessoa/delete?id=<?= $item->id ?>" ><button>EXCLUIR</button></a></td>
+                </tr>
+            <?php endforeach; ?>
+
+        </table>
+
+    <?php else: ?>
+
+        <h2 align="center"> Nenhum registro encontrado </h2>
+
+    <?php endif; ?> 
 
 </main>
